requirements.txt
streamlit==1.46.1
openai==1.101.0
python-dotenv
streamlit_pdf_viewer==0.0.26
PyMuPDF==1.26.3
sqlalchemy==2.0.43
pytest-check==2.5.3